import: Add options to control data-type and parallelism when importing a model
This change adds the ability to specify '--data-type' when running 'triton import'.
This value will be used to determine the data type used when compiling and optimizing the model for TRT-LLM.

This change adds the ability to specify '--pipeline-parallelism' and '--tensor-parallelism' when running 'triton import'.
This value will be used to determine the parallelism options used when compiling and optimizing the model for TRT-LLM.